feat: Update navigation logic in MaisLixo screen to use Login instead of Cadastro
This commit updates the navigation logic in the MaisLixo screen to use the 'Login' screen instead of the 'Cadastro' screen when the 'Junte-se a nós' button is pressed. This change ensures that users are directed to the correct screen for logging in, improving the user experience and aligning with the Firebase authentication implementation.